testing testing

+++ ParallelMergeSort/src/TestCase.java
@@ -2,8 +2,8 @@
  * Created by o_0 on 2017-03-14.
  */
 public class TestCase {
-    public static boolean inOrder(int[] arr) {
-        int lastValue = arr[0];
+    public static boolean inOrder(double[] arr) {
+        double lastValue = arr[0];
         for (int i = 1; i < arr.length; i++) {
             if (lastValue > arr[i]) {
                 return false;
